+++ lib/src/configuration/better_player_data_source.dart
@@ -38,11 +38,13 @@ class BetterPlayerDataSource {
   ///Optional cache configuration, used only for network data sources
   final BetterPlayerCacheConfiguration cacheConfiguration;
 
+  ///List of bytes, used only in memory player
   final List<int> bytes;
 
   BetterPlayerDataSource(
     this.type,
     this.url, {
+    this.bytes,
     this.subtitles,
     this.liveStream = false,
     this.headers,
@@ -51,18 +53,26 @@ class BetterPlayerDataSource {
     this.hlsTrackNames,
     this.resolutions,
     this.cacheConfiguration,
-    this.bytes,
-  });
+  }) : assert(
+            ((type == BetterPlayerDataSourceType.NETWORK ||
+                        type == BetterPlayerDataSourceType.FILE) &&
+                    url != null) ||
+                (type == BetterPlayerDataSourceType.MEMORY &&
+                    bytes?.isNotEmpty == true),
+            "Url can't be null in network or file data source | bytes can't be null when using memory data source");
 
+  ///Factory method to build network data source which uses url as data source
+  ///Bytes parameter is not used in this data source.
   factory BetterPlayerDataSource.network(
-      String url,
-      List<BetterPlayerSubtitlesSource> subtitles,
-      bool liveStream,
-      Map<String, String> headers,
-      bool useHlsSubtitles,
-      bool useHlsTracks,
-      Map<String, String> qualities,
-      BetterPlayerCacheConfiguration cacheConfiguration) {
+    String url, {
+    List<BetterPlayerSubtitlesSource> subtitles,
+    bool liveStream,
+    Map<String, String> headers,
+    bool useHlsSubtitles,
+    bool useHlsTracks,
+    Map<String, String> qualities,
+    BetterPlayerCacheConfiguration cacheConfiguration,
+  }) {
     return BetterPlayerDataSource(
       BetterPlayerDataSourceType.NETWORK,
       url,
@@ -75,14 +85,16 @@ class BetterPlayerDataSource {
       cacheConfiguration: cacheConfiguration,
     );
   }
-
+  ///Factory method to build file data source which uses url as data source.
+  ///Bytes parameter is not used in this data source.
   factory BetterPlayerDataSource.file(
-      String url,
-      List<BetterPlayerSubtitlesSource> subtitles,
-      bool useHlsSubtitles,
-      bool useHlsTracks,
-      Map<String, String> qualities,
-      BetterPlayerCacheConfiguration cacheConfiguration) {
+    String url, {
+    List<BetterPlayerSubtitlesSource> subtitles,
+    bool useHlsSubtitles,
+    bool useHlsTracks,
+    Map<String, String> qualities,
+    BetterPlayerCacheConfiguration cacheConfiguration,
+  }) {
     return BetterPlayerDataSource(
       BetterPlayerDataSourceType.NETWORK,
       url,
@@ -92,14 +104,14 @@ class BetterPlayerDataSource {
       resolutions: qualities,
     );
   }
-
-  factory BetterPlayerDataSource.memory(
-      List<int> bytes,
-      List<BetterPlayerSubtitlesSource> subtitles,
+  ///Factory method to build network data source which uses bytes as data source.
+  ///Url parameter is not used in this data source.
+  factory BetterPlayerDataSource.memory(List<int> bytes,
+      {List<BetterPlayerSubtitlesSource> subtitles,
       bool useHlsSubtitles,
       bool useHlsTracks,
       Map<String, String> qualities,
-      BetterPlayerCacheConfiguration cacheConfiguration) {
+      BetterPlayerCacheConfiguration cacheConfiguration}) {
     return BetterPlayerDataSource(
       BetterPlayerDataSourceType.MEMORY,
       "",